added road runner stuff

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/shapes/Point.java
@@ -1,8 +1,8 @@
 package org.firstinspires.ftc.teamcode.shapes;
 
 public class Point {
-    long x, y;
-    public Point(long _x, long _y) {
+    int x, y;
+    public Point(int _x, int _y) {
         this.x = _x;
         this.y = _y;
     }
